+++ docs/changelog.rst
@@ -13,6 +13,9 @@ CHANGELOG
 **Minor changes**
 
 * Improve HTML templates for public exports
+* Multiple tel numbers in SITRA parser
+* Import Adresse3 line in SITRA parser
+* Move descriptifCourt to description_teaser in SITRA parser
 
 **Bug fixes**